processor/requirements.txt
psycopg2-binary>=2.9.9
minio>=7.1.13
ebooklib>=0.18
beautifulsoup4>=4.12.2
lxml>=4.9.3
nltk>=3.8.1
python-dotenv>=1.0.0
